make new controller file, addFavs
