+++ src/server/routes/userRouter.js
@@ -1,12 +1,13 @@
 const express = require('express');
 const userController = require('../controllers/userController');
+const favoritesController = require('../controllers/favoritesController');
 const router = express.Router();
 
 router.post(
   '/',
   userController.verifyUser,
-  userController.getFavTechs,
-  userController.getFavResources,
+  favoritesController.getFavTechs,
+  favoritesController.getFavResources,
   (req, res) => {
     // send login confirmation and user info back to frontend
     return res.status(200).json({
@@ -20,23 +21,32 @@ router.post(
 router.post(
   '/favorite',
   userController.validateToken,
-  userController.checkFavResource,
-  userController.addFavResource,
+  favoritesController.checkFavResource,
+  favoritesController.addFavResource,
+  favoritesController.getFavResources,
   (req, res) => {
-    return res.status(200).json(res.locals.users_id);
+    return res.status(200).json({ favoriteResources: res.locals.favResources });
   }
 );
 
 router.delete(
   '/favorite',
   userController.validateToken,
-  userController.removeFavResource,
+  favoritesController.removeFavResource,
+  favoritesController.getFavResources,
   (req, res) => {
-    return res.status(200).end();
+    return res.status(200).json({ favoriteResources: res.locals.favResources });
   }
 );
 
-// router.get('/favorite', userController.validateToken, etc..)
 // get all favorite resources from this particular user
+router.get(
+  '/favorite',
+  userController.validateToken,
+  favoritesController.getFavResources,
+  (req, res) => {
+    return res.status(200).json({ favoriteResources: res.locals.favResources });
+  }
+);
 
 module.exports = router;
